12 read xml labels (#13)
* Add Binding Profile Doc

Add binding profile documentation to the project.
Updated version number to 1.1.0.

* Add Move Semantics

Add move ctor and assignment operator to the AccessUnit class.

* Handle Label Format

Handle when the label is encoded as EXI or XML.

* Bump Version

Bump Version number to 1.2.0.

* Callbacks

Using callbacks and lambda function to implement a callback function when the LabelDemux extracts a label.

* Minor Change

Add a new return before printing the results.

* Add New Test Case

Added a test case if the transport stream contains labels that are XML text-formated.
Updated comments in LabelDemux class.

* Workaround

There is bug in ConfBind when packetizing an XML label where it adds extra NULL characters.   I added a workaround so that the XML label is streamed as a C null terminated string instead as a C++ string.