Handle Bad PAT (#16)
In LabelDemux, update the PAT instance every time it reads one in the transport stream.  Before, the LabelDemux would only read the first one.  Therefore, if the first PAT was malformed, the LabelDemux would fail to find the PMT.

Bump version number to 1.2.1.

In main, if no output file is specified, ensure to prints the labels out to console.